+++ CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+- Endpoints to get practice codes and names for EQIP and CSP. [#263](https://github.com/policy-design-lab/pdl-api/issues/263)
+
+## [0.19.0] - 2024-11-03
+
 ### Changed
 - Landing page values to use data from the database. [#258](https://github.com/policy-design-lab/pdl-api/issues/258)
+
+### Fixed
+- Title-II total payment calculation and overall benefits payment amounts calculation in the GET /pdl/allprograms endpoint. [#268](https://github.com/policy-design-lab/pdl-api/issues/268) 
 
 ## [0.18.0] - 2024-10-07
 
@@ -190,6 +199,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Open API specification file. [#50](https://github.com/policy-design-lab/pdl-api/issues/50)
 
+[0.19.0]: https://github.com/policy-design-lab/pdl-api/compare/0.18.0...0.19.0
 [0.18.0]: https://github.com/policy-design-lab/pdl-api/compare/0.17.0...0.18.0
 [0.17.0]: https://github.com/policy-design-lab/pdl-api/compare/0.16.0...0.17.0
 [0.16.0]: https://github.com/policy-design-lab/pdl-api/compare/0.15.0...0.16.0